src/requirements.txt
absl-py==1.4.0
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
array-record==0.4.0
astunparse==1.6.3
asyncpg==0.28.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
dm-tree==0.1.8
etils==1.3.0
fastapi==0.100.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-resources==5.12.0
keras==2.13.1
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
promise==2.3
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.0.2
pydantic_core==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-hub==0.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.1
termcolor==2.3.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
